Massive changes made to the functionality
redone the way it shows content entirely :
removed table element from HTML and all of its javascript functionality.
instead added bootstrap cards to show the img, title and text of the show also added grid system to make it look better. All elements divs and classes are created and then appended inside of js. Added a lot of new variables. Also made changes to the search input field it is now in a separate container  in order to  round edges and add a magnifying icon to the left side, made corresponding changes to the button as well